refactor(extension-loader): split extension-loader.ts into per-concern submodules
Replace agent/extension-loader.ts (657 LOC) with an extension-loader/
directory grouping the four extension sources + theme loading:

- shared.ts    - ExtensionLoaderDeps + LoadHooks interfaces
- themes.ts    - RESERVED_THEME_IDS, normalizeTheme,
                 loadAethonThemeDirectory (loose-file themes from
                 ~/.aethon/themes/*.json)
- directory.ts - loadAethonExtensionDirectory (generic loader) +
                 loadAethonExtensions (user-dir wrapper) +
                 loadProjectAethonExtensions (project-dir wrapper) +
                 projectExtensionDisplayName
- packages.ts  - loadAethonExtensionPackages (npm-style packages
                 under ~/.aethon/skills/node_modules/, with scoped
                 namespace recursion + frontendEntry shipping)
- discovery.ts - discoverPiAethonExtensions (grep ~/.pi/agent/
                 extensions for globalThis.aethon markers) +
                 discoverPersistedTabs (sessions dir walk)
- index.ts     - barrel re-exporting the full public surface

Public surface unchanged. All 4 external consumers (main.ts,
aethon-api.ts, projectLifecycle.ts, tabs.ts) plus
extension-loader.test.ts keep importing from "./extension-loader"
which now resolves to the directory's index.ts. The
state.currentExtensionLoadScope / currentExtensionName register-on-stack
pair stays in directory.ts where it's set by register(api) calls.
